v0.9.0 public release

+++ Src/Public/Invoke-AsBuiltReport.Microsoft.AD.ps1
@@ -5,7 +5,7 @@ function Invoke-AsBuiltReport.Microsoft.AD {
     .DESCRIPTION
         Documents the configuration of Microsoft AD in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        0.8.1
+        Version:        0.9.0
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -21,10 +21,20 @@ function Invoke-AsBuiltReport.Microsoft.AD {
         [PSCredential] $Credential
     )
 
-    Write-PScriboMessage -IsWarning "Please refer to the AsBuiltReport.Microsoft.AD github website for more detailed information about this project."
-    Write-PScriboMessage -IsWarning "Do not forget to update your report configuration file after each new release."
-    Write-PScriboMessage -IsWarning "Documentation: https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.AD"
-    Write-PScriboMessage -IsWarning "Issues or bug reporting: https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.AD/issues"
+    #Requires -Version 5.1
+    #Requires -PSEdition Desktop
+    #Requires -RunAsAdministrator
+
+    if ($psISE) {
+        Write-Error -Message "You cannot run this script inside the PowerShell ISE. Please execute it from the PowerShell Command Window."
+        break
+    }
+
+    Write-PScriboMessage -Plugin "Module" -IsWarning "Please refer to the AsBuiltReport.Microsoft.AD github website for more detailed information about this project."
+    Write-PScriboMessage -Plugin "Module" -IsWarning "Do not forget to update your report configuration file after each new release."
+    Write-PScriboMessage -Plugin "Module" -IsWarning "Documentation: https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.AD"
+    Write-PScriboMessage -Plugin "Module" -IsWarning "Issues or bug reporting: https://github.com/AsBuiltReport/AsBuiltReport.Microsoft.AD/issues"
+    Write-PScriboMessage -Plugin "Module" -IsWarning "This project is community maintained and has no sponsorship from Microsoft, its employees or any of its affiliates."
 
     Try {
         $InstalledVersion = Get-Module -ListAvailable -Name AsBuiltReport.Microsoft.AD -ErrorAction SilentlyContinue | Sort-Object -Property Version -Descending | Select-Object -First 1 -ExpandProperty Version
@@ -40,14 +50,6 @@ function Invoke-AsBuiltReport.Microsoft.AD {
     } Catch {
         Write-PScriboMessage -IsWarning $_.Exception.Message
     }
-
-    $currentPrincipal = New-Object Security.Principal.WindowsPrincipal([Security.Principal.WindowsIdentity]::GetCurrent())
-
-    if (-not $currentPrincipal.IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)) {
-
-        throw "The requested operation requires elevation: Run PowerShell console as administrator"
-    }
-
 
     #Validate Required Modules and Features
     $OSType = (Get-ComputerInfo).OsProductType
@@ -87,13 +89,29 @@ function Invoke-AsBuiltReport.Microsoft.AD {
     #---------------------------------------------------------------------------------------------#
     foreach ($System in $Target) {
 
+        if (Select-String -InputObject $System -Pattern "^\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}$") {
+            throw "Please use the FQDN instead of an IP address to connect to the Domain Controller: $System"
+        }
+
         Try {
-            Write-PScriboMessage "Connecting to Domain Controller Server '$System'."
+            Write-PScriboMessage "Connecting to Domain Controller through PSSession $System"
             $script:TempPssSession = New-PSSession $System -Credential $Credential -Authentication $Options.PSDefaultAuthentication -ErrorAction Stop -Name "Global:TempPssSession"
-            $script:TempCIMSession = New-CimSession $System -Credential $Credential -Authentication $Options.PSDefaultAuthentication -ErrorAction Stop -Name "Global:TempCIMSession"
+        } Catch {
+            throw "Unable to connect to the Domain Controller through PSSession: $System"
+        }
+
+        Try {
+            Write-PScriboMessage "Connecting to Domain Controller through CimSession '$System'."
+            $script:TempCIMSession = New-CimSession $System -Credential $Credential -Authentication $Options.PSDefaultAuthentication -ErrorAction Continue -Name "Global:TempCIMSession"
+        } Catch {
+            Write-PScriboMessage -IsWarning "Unable to connect to the Domain Controller through CimSession: $System"
+        }
+
+        Try {
+            Write-PScriboMessage "Connecting to get Forest information from Domain Controller '$System'."
             $script:ADSystem = Invoke-Command -Session $TempPssSession { Get-ADForest -ErrorAction Stop }
         } Catch {
-            throw "Unable to connect to the Domain Controller: $System"
+            throw "Unable to get Forest information from Domain Controller: $System"
         }
 
         $script:ForestInfo = $ADSystem.RootDomain.toUpper()
@@ -113,13 +131,18 @@ function Invoke-AsBuiltReport.Microsoft.AD {
         # PKI Section
         Get-AbrPKISection
 
-        # Remove used PSSession
-        Write-PScriboMessage "Clearing PowerShell Session $($TempPssSession.Id)"
-        Remove-PSSession -Session $TempPssSession
+        if ($TempPssSession) {
+            # Remove used PSSession
+            Write-PScriboMessage "Clearing PowerShell Session $($TempPssSession.Id)"
+            Remove-PSSession -Session $TempPssSession
+        }
 
-        # Remove used CIMSession
-        Write-PScriboMessage "Clearing CIM Session $($TempCIMSession.Id)"
-        Remove-CimSession -CimSession $TempCIMSession
+        if ($TempCIMSession) {
+            # Remove used CIMSession
+            Write-PScriboMessage "Clearing CIM Session $($TempCIMSession.Id)"
+            Remove-CimSession -CimSession $TempCIMSession
+        }
+
 
     }#endregion foreach loop
 }
